requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.27.1
selenium>=4.1.0
webdriver-manager>=3.5.2
beautifulsoup4>=4.10.0
lxml>=4.6.3

# CLI and utils
argparse>=1.4.0
colorama>=0.4.4

# Development tools
pytest>=7.0.0
flake8>=4.0.1
black>=22.1.0

# Type checking
mypy>=0.931
types-requests>=2.27.11

